Remove VoiceStitcher, consolidate into ChirpLinker
Delete voice.py — move DAG-building, greedy assignment, and config
into legato.py. Rename VoiceStitchConfig → ChirpLinkConfig. Update
all demos to use ChirpLinker exclusively.

The anchor-point voice representation is removed in favor of enriched
tokens (same tensor shape, directly feedable to transformers).
