+++ src/fuge/__init__.py
@@ -1,7 +1,7 @@
 """fuge: Scientific signal embeddings.
 
 Subpackages:
-    fuge.spectral  — Chirp tokenization and voice stitching (ChirpTokenizer, VoiceStitcher, ChirpTokenEmbedding, DechirpSTFT)
+    fuge.spectral  — Chirp tokenization and linking (ChirpTokenizer, ChirpLinker, ChirpTokenEmbedding, DechirpSTFT)
     fuge.svd       — Streaming PCA with Procrustes stabilization (StreamingPCA)
     fuge.nn        — Neural network components (TransformerEmbedding)
 """
