+++ src/fuge/spectral/legato.py
@@ -13,9 +13,131 @@ field (chain_id), making it directly usable by downstream transformers.
 
 import torch
 import torch.nn as nn
+from dataclasses import dataclass
 
 from fuge.spectral.tokens import ChirpTokens, N_BASE
-from fuge.spectral.voice import VoiceStitchConfig, _build_dag, _greedy_assign, _wrap
+
+
+@dataclass
+class ChirpLinkConfig:
+    """Matching thresholds for token linking.
+
+    Parameters
+    ----------
+    max_df : float
+        Max relative frequency mismatch |f_end[w] - f_start[w+1]| / f_mean.
+    max_dphi : float
+        Max boundary phase residual |wrap(φ_start[w+1] - φ_end[w])| in radians.
+    max_dA : float
+        Max relative amplitude mismatch |A_end[w] - A_start[w+1]| / max(A).
+    """
+    max_df: float = 0.05
+    max_dphi: float = 0.5
+    max_dA: float = 0.5
+
+
+def _wrap(x: torch.Tensor) -> torch.Tensor:
+    """Wrap angle to [-π, π]."""
+    return (x + torch.pi) % (2 * torch.pi) - torch.pi
+
+
+def _build_dag(tokens: torch.Tensor, cfg: ChirpLinkConfig
+               ) -> tuple[list[list[tuple[int, int]]], dict]:
+    """Build DAG of compatible tokens and resolve into chains.
+
+    Parameters
+    ----------
+    tokens : (W, K, C) tensor for a single batch element (C >= 9).
+    cfg : matching thresholds.
+
+    Returns
+    -------
+    edges : list of list of (k_prev, k_next) per window boundary.
+    chains : dict mapping (w, k) -> (total_snr, [path]).
+    """
+    W, K = tokens.shape[0], tokens.shape[1]
+
+    snr = tokens[..., 0]
+    f_start = tokens[..., 3]
+    f_end = tokens[..., 4]
+    A_start = tokens[..., 5]
+    A_end = tokens[..., 6]
+    ps = tokens[..., 7]
+    pe = tokens[..., 8]
+
+    edges = []
+    for w in range(W - 1):
+        window_edges = []
+        for kp in range(K):
+            if snr[w, kp] <= 0:
+                continue
+            for kn in range(K):
+                if snr[w + 1, kn] <= 0:
+                    continue
+                fe = f_end[w, kp]
+                fs = f_start[w + 1, kn]
+                f_mean = (fe + fs) / 2
+                if f_mean > 0 and abs(fe - fs) / f_mean > cfg.max_df:
+                    continue
+                dphi = _wrap(ps[w + 1, kn] - pe[w, kp])
+                if abs(dphi) > cfg.max_dphi:
+                    continue
+                ae = A_end[w, kp]
+                an = A_start[w + 1, kn]
+                a_max = max(ae, an)
+                if a_max > 0 and abs(ae - an) / a_max > cfg.max_dA:
+                    continue
+                window_edges.append((kp, kn))
+        edges.append(window_edges)
+
+    # DP: best chain ending at each (w, k).
+    chains: dict[tuple[int, int], tuple[float, list[int]]] = {}
+
+    for k_idx in range(K):
+        if snr[0, k_idx] > 0:
+            chains[(0, k_idx)] = (snr[0, k_idx].item(), [k_idx])
+
+    for w in range(W - 1):
+        for kp, kn in edges[w]:
+            if (w, kp) not in chains:
+                continue
+            prev_snr, prev_path = chains[(w, kp)]
+            new_snr = prev_snr + snr[w + 1, kn].item()
+            if (w + 1, kn) not in chains or chains[(w + 1, kn)][0] < new_snr:
+                chains[(w + 1, kn)] = (new_snr, prev_path + [kn])
+
+    # Seed unreached tokens.
+    for w in range(1, W):
+        for k_idx in range(K):
+            if snr[w, k_idx] > 0 and (w, k_idx) not in chains:
+                chains[(w, k_idx)] = (snr[w, k_idx].item(), [k_idx])
+
+    return edges, chains
+
+
+def _greedy_assign(chains: dict, W: int) -> list[tuple[int, list[int]]]:
+    """Greedily assign non-overlapping chains by total SNR.
+
+    Returns list of (w_start, path) tuples.
+    """
+    all_chains = []
+    for (w, k_idx), (total_snr, path) in chains.items():
+        w_start = w - len(path) + 1
+        all_chains.append((total_snr, w_start, path))
+
+    all_chains.sort(key=lambda x: x[0], reverse=True)
+
+    used = set()
+    result = []
+    for total_snr, w_start, path in all_chains:
+        slots = [(w_start + i, path[i]) for i in range(len(path))]
+        if any(s in used for s in slots):
+            continue
+        for s in slots:
+            used.add(s)
+        result.append((w_start, path))
+
+    return result
 
 
 class ChirpLinker(nn.Module):
@@ -23,17 +145,17 @@ class ChirpLinker(nn.Module):
 
     Parameters
     ----------
-    config : VoiceStitchConfig or None
+    config : ChirpLinkConfig or None
         Matching thresholds (max_df, max_dphi, max_dA).
     min_length : int
         Minimum chain length to enrich. Shorter chains are left as-is
         with chain_id = -1.
     """
 
-    def __init__(self, config: VoiceStitchConfig | None = None,
+    def __init__(self, config: ChirpLinkConfig | None = None,
                  min_length: int = 2):
         super().__init__()
-        self.config = config or VoiceStitchConfig()
+        self.config = config or ChirpLinkConfig()
         self.min_length = min_length
 
     @torch.no_grad()
